[amqplib] do not pollute global namespace + cleanup to tests
